ClayScore : audit par des relecteurs EXTERIEURS + correctifs prouves
Kevin : « fais analyser par des avis exterieur, l'ensemble ».
6 analyseurs deterministes, aucun n'est une IA, aucun n'a de raison de
menager ce code : mypy, bandit, pip-audit, vulture, radon, pylint.
Plus la Security Suite en CI (gitleaks/TruffleHog/OSV/Trivy/Semgrep/zizmor).

DEUX VRAIS DEFAUTS, trouves par bandit, invisibles aux 341 tests :

1. shell=True (severite HAUTE, CWE-78) dans tools/build_pack.py
   Les chemins sont des constantes -> rien n'etait attaquable aujourd'hui,
   mais un dossier de travail avec une espace suffisait a casser la
   construction du dossier. Remplace par shutil.copytree : plus de shell.

2. urlopen acceptait file:// (CWE-22) dans tools/verif_prix.py
   Une URL de fournisseur mal saisie aurait fait LIRE un fichier de la
   machine, dont le contenu serait remonte dans le rapport de prix.
   Filtre http/https + 3 tests qui prouvent qu'un faux mot de passe pose
   dans un fichier ne fuit pas.

mypy : 8 erreurs -> 0. Aucune n'etait une panne en cours (verifie dans le
code avant d'y toucher, et dit comme tel), mais elles empechaient
d'utiliser mypy comme garde-fou, donc elles cachaient les vraies.
Dont __exit__ -> bool, qui laissait croire que le gestionnaire pouvait
AVALER une exception, et int(cur.lastrowid) qui plantait si aucun
identifiant n'etait rendu.

PREVENTION (le correctif qui compte) : tools/audit_externe.sh --gate
bloque si ruff/mypy/bandit-haute/pip-audit/tests echouent. Sans --gate il
ajoute couverture, code mort et complexite. `pip install -e ".[audit]"`.

Mesure : 344 tests verts (+3), couverture 91 %, ruff propre, mypy propre,
bandit 0 en severite haute, pip-audit 0 faille, vulture 0 code mort,
pylint 8,39/10. Suite rejouee depuis l'interieur du ZIP livre.

Le point le PLUS grave n'a ete trouve par aucun outil et est consigne
dans docs/AUDIT_EXTERIEUR.md § 2 : le depot est PUBLIC.

+++ clayscore/clayscore/calibration/corridor.py
@@ -74,7 +74,7 @@ class CorridorCalibrator:
         envelope = dev.std(axis=0) + dev.mean(axis=0) * 0.0  # base = std inter-essais
         # Plancher d'enveloppe pour éviter des tolérances nulles (bruit).
         scale = np.linalg.norm(reference.max(axis=0) - reference.min(axis=0))
-        floor = max(self.min_envelope_frac * scale, 1e-3)
+        floor = max(float(self.min_envelope_frac * scale), 1e-3)
         envelope = np.maximum(envelope, floor)
         return Corridor(reference=reference, envelope=envelope,
                         n_samples=len(tracks), dims=reference.shape[1])
